requirements.txt
kafka-python
requests
confuse~=1.0.0
